Organize classes, add LocalValidationRunner
Because:
- keeps concerns organized and clear
- new code is being added

* LocalValidationRunner is logic for internal use
* Rename Validator to Validation in several classes

Ticket:#579
