+++ src/main/java/org/pdxfinder/validator/LocalValidationRunner.java
@@ -1,0 +1,12 @@
+package org.pdxfinder.validator;
+
+import org.springframework.boot.CommandLineRunner;
+
+public class LocalValidationRunner implements CommandLineRunner {
+
+  @Override
+  public void run(String... arg) throws Exception {
+    if (CommandCli.inLocalMode(arg)) {
+    }
+  }
+}
