+++ src/main/java/org/pdxfinder/validator/api/Uploader.java
@@ -1,7 +1,6 @@
 package org.pdxfinder.validator.api;
 
 import java.util.Optional;
-import org.pdxfinder.validator.ValidatorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,16 +10,17 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+
 @RestController
 @RequestMapping("/validation/")
 public class Uploader {
 
-  private ValidatorService validatorService;
+  private ValidationWebService validatorService;
   private static final String EXCEL_CONTENT_TYPE =
       "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
 
   @Autowired
-  Uploader(ValidatorService validatorService) {
+  Uploader(ValidationWebService validatorService) {
     this.validatorService = validatorService;
   }
